# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rayito de Sol API",
        "description": "Documentación de la API para Rayito de Sol",
        "contact": {
            "email": "sergioramosalarcon@rayitodesol.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/properties/{propertyId}/day-prices": {
            "get": {
                "tags": [
                    "Propiedades"
                ],
                "summary": "Obtener los precios diarios personalizados del mes para una propiedad",
                "operationId": "5689a403807ec0c26517faae74fb1689",
                "parameters": [
                    {
                        "name": "propertyId",
                        "in": "path",
                        "description": "ID de la propiedad",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Año (YYYY). Por defecto el año actual.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2025
                        }
                    },
                    {
                        "name": "month",
                        "in": "query",
                        "description": "Mes (1-12). Por defecto el mes actual.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 7
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Precios diarios del mes consultado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "2025-07-01": {
                                            "properties": {
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 120
                                                },
                                                "notes": {
                                                    "type": "string",
                                                    "example": "Precio especial por evento"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Propiedades"
                ],
                "summary": "Crear o actualizar el precio personalizado para un día concreto en una propiedad",
                "operationId": "ae662d282215e9fbc07f8bf7e6e218a8",
                "parameters": [
                    {
                        "name": "propertyId",
                        "in": "path",
                        "description": "ID de la propiedad",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "price"
                                ],
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-01"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Precio por temporada alta"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Precio guardado o actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Propiedades"
                ],
                "summary": "Eliminar el precio personalizado de un día concreto en una propiedad",
                "operationId": "3734f117ea272db3563cae100da31647",
                "parameters": [
                    {
                        "name": "propertyId",
                        "in": "path",
                        "description": "ID de la propiedad",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date"
                                ],
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Precio eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}